generate fatal error when main output isn't a Scene
Up to now, the player didn't show anything if the
output of the Home or loader progress operators
weren't of type "Scene", not even the slightest
warning.

This commit makes this a fatal error: the player
will quit immediately and leave a meaningful
message in the log files.

(This may seem trivial, but if you spent two
hours debugging "Player doesn't work" issues
on a party with a looming deadline, you'd
really appreciate this :)